Allow Head Map Dev to manage Develop subforum
I believe that this subforum can be used for more than reporting bugs and some ideas. I was thinking of being able to have head map developers involved with more cooperation towards sectors and have it serve as for open discussions on certain topics. As head map developers are also senior staff and work closely with developers, I see it as appropriate if head map developers can serve in such ways.